lambda_boilerplate/pyproject.toml
[project]
name = "lambda-boilerplate"
version = "0.1.0"
description = "A production-ready AWS Lambda boilerplate with structured logging."
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
